Update README.md
NoteVault is a collaborative web app that allows students to upload and retrieve any kind of notes. The project is developed with an intention to reduce the efforts of students wasted in finding academic resources. Coordinators can verify the legitimacy of the resource and make them public. Students can browse through the resources and upload any useful material if they wish.